refactor(tests): 重构测试代码并添加新方法
- 移除了大量的模拟实现代码，简化了测试结构
- 添加了 MockTaskControlService 的新方法：
  - cancel_all_task_runs
  - has_running_instances
  - get_recent_executions
- 更新了测试用例，使用新的方法进行测试
- 删除了一些不必要的测试用例

+++ crates/api/src/lib.rs
@@ -437,7 +437,7 @@ mod tests {
         async fn get_recent_executions(
             &self,
             _task_id: i64,
-            _limit: i64,
+            _limit: usize,
         ) -> scheduler_core::SchedulerResult<Vec<scheduler_domain::entities::TaskRun>> {
             unimplemented!()
         }
